Move stack to provider module
Moves `stack`, `stack_status` and `stack_policies` to `provider`
