+++ sceptre/resolvers/__init__.py
@@ -28,7 +28,7 @@ class Resolver(ResolverData):
     :param argument: The argument of the resolver.
     :type argument: str
     :param stack: The associated stack of the resolver.
-    :type stack: sceptre.stack.Stack
+    :type stack: sceptre.provider.stack.Stack
     """
 
     __metaclass__ = abc.ABCMeta
